build: esm and cjs (#9)
# Motivation

Bundle the library to ES modules and Common JS in order to be able to use it in nns-dapp

# Changes

- bundle the library with esbuild
- generate esm modules with code splitting
- provide type declarations
- exclude `test.ts` and `spec.ts` from bundle
- fix test error `unexpected token 'export'`

# isolatedModules issue with reused agent-js const enum

Using a compiler (esbuild, rollup, etc.) leads to the same restriction as enabling `isolatedModules` for TypeScript to ensure that the compiled code is safe in case `const enum` would be exported (see [blog post](https://ncjamieson.com/dont-export-const-enums/)).

While developing this PR I noticed that the current code does lead to this behavior. Therefor I replaced the use of the enum with a plain `string`. This should be improved.

# Note

Same goal as PR [#3](https://github.com/dfinity/nns-js/pull/3) but produces split chunks of ES modules, generates types declarations, maybee has a cleaner bundle structure and use a modern bundler.

Usage tested in nns-dapp ✅.

Let's discuss it.