Update governance.did (#19)
# Motivation
The candid file in nns-js used to contain only a small subset of the upstream governance did file.  We will shortly need most, if not all, of the API.

# Changes
- Import the governance candid did file from the ic repo.
- Compile idl.js from this file.

# Testing
Note that MOST changes just add methods, so should be non-breaking.  There is only one breaking change:  Service now takes an argument.  It is this that needs the most careful testing.