List neurons (#33)
# Motivation
We need to list a user's neurons in the svelte nns dapp.

# Changes
Adds the `getNeurons()` method, that gets a caller's own neurons.  The method mirrors the `getNeurons` in `nns-dapp/frontend/ts` however I am a little concerned by the API and would be very open to exposing the same API as raw.