Feat: Add createAt param for NNS Transfers (#255)
# Motivation

Add optional param createdAt in NNS Transfers.

This parameter is used in the NNS Ledger to identify duplicated
transactions. Without it, two transfers with the same params from the
same caller would be considered as two different transactions. This
could happen if agent-js or the client retries the request.

If the request has the createdAt set and agent-js retries the call, the
Ledger canister will consider it as a duplicated and not perform the
transaction twice.

# Changes

* Add createdAt optional param to `TransferRequest` type.
* Handle the new param in transferHardwareWallet.
* Handle the new param in toTransferRawRequest.

# Tests

* Add a test to check that the new param is added in the request.