Add ICRC-1 version of stakeNeuron method (#418)
# Motivation

We want use ICRC-1 instead of generating account identifiers in the
client.
When staking a neuron we first have to transfer the stake to a
subaccount of the governance canister.
We want to use an ICRC-1 transfer to do this.

# Changes

Add `stakeNeuronIcrc1` which is a copy of `stakeNeuron` with minimal
changes to use `IcrcLedgerCanister` instead of `LedgerCanister`. The
plan is for this to replace `stakeNeuron` once it's been properly tested
in nns-dapp.

# Tests

Add a copy of all the `stakeNeuron` tests and apply minimal changes to
make them apply to the new `stakeNeuronIcrc1` method.

Tested manually by installing the local package in a local nns-dapp and
using the new method to create a neuron. I tested increasing stake,
setting dissolve delay and voting with this new neuron.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>