build: init new library ledger-icrc-js (#428)
# Motivation

This PR initiate a new library `@dfinity/ledger-icrc` which will replace `@dfinity/ledger` and in which all features will be moved.

I thought that creating a PR that initate the library and another PR in which we moved the code is a bit clear than a big fat rename. No strong opinion though we can also proceed that way or other way, open to better ideas.