build: publish to npm only if needed (#432)
# Motivation

So far, we've been incrementing the version of all libraries when we publish, even if some of them received no improvements or fixes. I find this approach a bit cumbersome. Since I propose to commit to semantic versioning (#431), a better approach would be most welcomed. That's why this PR introduces a publication script that handles the publishing of libraries only if they were actually modified.

# Changes

- use `shasum` to compare local library with npm and publish only upon differences