feat: copy ICP ledger-related features to new ledger-icp library (#427)
# Motivation

Community and developers are confused about what ledger and interfaces to use in which context. That's why we decided to makes it more clear that interfacing with ICP requires a specific library even though it supports ICRC-1. The idea is to have a generic ICRC-1 ledger lib and one dedicated to ICP.

That's why we are creating a new library and moving all ledger ICP related code from `@dfinity/nns` to `@dfinity/ledger-icp`.

# Notes

- No code changes, only code copied around.
- In class `AccountIdentifier` instead of using a type declared in the Governance candid file to returns a hash as Unit8Array, I used an object that contains the primitive type as declaration.
- This PR copy the related content to `ledger-icp` and a future PR will remove the duplication in `nns`
- We also explicitely not copy the deprecated `ICP` to definitely ultimately remove it