Chore: Change "?" for "Option" for optional fields (#480)
# Motivation

Using `?` is not the same as `Option`. With `?` the field is optional,
but with `Option`, the field is mandatory, but the value can be
`undefined`.

This small distinction is important when we render the information of
proposals in NNS Dapp. We convert the actions to camelcase and remove
`[] | [T]`.

If we use `?`, we might forget to convert it and it will never show up
in NNS Dapp. By using `Option`, we make sure that the field will always
we present in the converters.

You can see the example in the `params`, `maxDirectParticipationIcpE8s`
and `minDirectParticipationIcpE8s` fields for the (deprecated)
`OpenSnsTokenSwap` proposal action that I had to change when I moved
from `?` to `Option`.

# Changes

* Change the `?` in `types/governance_converters.ts` for `Option`. Only
for the fields related to proposals.

# Tests

* Fix a converter after changes.

# Todos

- [x] Add entry to changelog (if necessary).