Fix: Support converting small number to AmountTokenV2 (#496)
# Motivation

TokenAmountV2 can't convert small numbers such as `0.00000002` which is
2 e8s of ICP.

The problem was that `fromNumber` was using `toString` and `toString`
was converting `0.00000002` to the scientific notation. Which we
specifically don't want to support when converting from a string.

Solution, use `toFixed(8)` instead of `toString`. We already support
only a precision of 8 decimals.

# Changes

* Use `toFixed(8)` instead of `toString` in `fromNumber` of
TokenAmountV2.

# Tests

* Change test `"does not support scientific notation"` to do it
`fromString`.
* Add tests for 8 decimals.
* Add a test for `fromNumber` with very small numbers.
* Add test for `fromNumber` to check that precision is 8 decimals.

# Todos

- [x] Add entry to changelog (if necessary).

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>