fix: iOS 15 usage and bump version (#513)
# Motivation

It was reported that the last version of NNS dapp shipped with SvelteKit v2 and per extension Rollup v4 was not usable on iOS 15 (iOS 17 was fine). Afternarrowing down the issue I figured out that the root cause of the issue was the usage of a static variable using within an object parameter.

# Error stacktrace on iOS v15

```
Unhandled Promise Rejection: SyntaxError: Unexpected token '{'
```

# Changes

- Replace static usage of `SubAccount.ZERO` with an object to fix the issue
- Add entry in CHANGELOG
- Bump version to release the patch to npm asap.

# Screenshot

![image](https://github.com/dfinity/ic-js/assets/16886711/f6581810-d0de-4f74-bcbf-3133fb31f112)

# Test

The fix was tested locally by bundling the library and including it in NNS dapp which was build and served `npx serve public` + `npx local-ssl-proxy --source 3001 --target 3000`. Finally the fix was validated by loading https://locahost:3001 in Xcode simulator with iOS 15.4.