build: upgrade didc v0.3.7 (#603)
# Motivation

I've been using didc v0.3.7 for some time and I'm tired of having to switch tools just for this repository. Therefore, this PR bumps didc to v0.3.7.

Compared to v0.3.5, didc v0.3.7 generates nicer service interfaces. Interfaces which are for example useful when one install a canister in JS as they can import the `init` types.

# Changes

- bump didc v0.3.7 in GitHub Actions
- set min required version to 0.3.7