feat: ckETH minter info update for ckERC20 (#578)
# Motivation

There are new fields provided in the minter info of ckETH (source #606) and there is also one that was renamed, therefore a breaking changes (field we do not use yet in Oisy btw.).

# Notes

From the minter info, in Oisy we are currently using `minimum_withdrawal_amount` and `last_observed_block_number` which are not modified in the did file. I've also tested locally Oisy with old/new version of the library against old/new version of the canisters. All combination are functional. Also tested Oisy with the next version on staging against the deployed testnet canister, it was functional too.

# Changes

- Bump did
- Add a notes about breaking changes in CHANGELOG
- Update related test mock data