feat: Bitcoin canister get utxos (#644)
# Motivation

The `get_utxos` functions of `@dfinity/ic-management` called the Bitcoin canister under the hood. This PR basically copies those two functions (update and query) to the new Bitcoin canister with the difference that instead of exposing two functions, the JS library will expose a single function that can be called with a parameter `certified`. This aligns our general strategy.

# Changes

- Copy function `bitcoinGetUtxos` and `bitcoinGetUtxosQuery` from IC-mgmt to Bitcoin canister
- Adapt did
- Merge into a single function which accepts query or update call
- Rename copied content to match naming of new canister