asNonNullish like assertNonNullish but returns the value (#675)
# Motivation

We sometimes have a value of type `T | undefined` but we know it is of
type `T` so we use `foo as T` before using it.
It would be safer to use `assertNonNullish` first but this can be
tedious because it can't be used inline.

For example, something as simple as:
```
const getToken = (universe) => {
  return universe.canisterId == OWN_CANISTER_ID_TEXT ? ICPToken : universe.summary?.token as Token;
};
```
would become
```
const getToken = (universe) => {
  if (universe.canisterId == OWN_CANISTER_ID_TEXT) {
    return ICPToken;
  } else {
    assertNonNullish(universe.summary);
    return universe.summary.token;
  }
};
```
It would be nicer if we could use `assertNonNullish` inline like this:
```
const getToken = (universe) => {
  return universe.canisterId == OWN_CANISTER_ID_TEXT ? ICPToken : assertNonNullish(universe.summary).token;
};
```

I've tried to make this work but it doesn't seem possible to both assert
the type of the argument and also return the argument. See
https://stackoverflow.com/questions/78770654/typescript-assertion-function-that-returns-asserted-value

So I'm adding an extra function `asNonNullish` which can be used as in
the second use case.

# Changes

1. Added `asNonNullish` which asserts that its argument is non-nullish
and returns it.
2. Simply the syntax of `assertNonNullish` by using `function` instead
of `=>`.

# Tests

1. Added a missing unit test for `assertNonNullish` which tests (at
compile time) that the type of the value becomes non-nullish. This test
unfortunately doesn't pass if `assertNonNullish` is replaced with
`asNonNullish`.
3. Added the same unit tests for `asNonNullish` that already existed for
`assertNonNullish`.
4. Added one more unit test for `asNonNullish` that tests that the value
is returned. This test would not pass for `assertNonNullish`.

# Todos

- [x] Add entry to changelog (if necessary).

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>