feat: bitcoin balance and getUtxos support for queries only (#706)
# Motivation

As we learned this week, update calls to the Bitcoin canister for
`get_balance` and `get_utxos` are only possible when executed from a
canister. Therefore, we remove the support for certified calls regarding
those endpoints in `@dfinity/ckbtc`. In addition, to make the exception
more obvious and to reflects the actual IC API naming, we also add the
suffix `Query` to the functions.

# Documentation


https://internetcomputer.org/docs/current/references/ic-interface-spec/#ic-bitcoin_get_utxos

# Changes

- Remove `certified` option parameter for `getBalance` and `getUtxos`
- Rename `getBalance` and `getUtxos` to `getBalanceQuery` and
`getUtxosQuery`
- Add a warning in the documentation
- Update code to use only endpoints `_query`
- Add breaking change in changelog

---------

Signed-off-by: David Dal Busco <david.dalbusco@dfinity.org>
Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>