feat: expose ledger-icp converters (#716)
# Motivation

The ICRC signer standards require developers to pass the payload as a
Candid `Uint8Array`. Such an array can be generated using
`@dfinity/candid` which I find it somewhat cumbersome to use when you
are not used to it. Moreover, the IDL declarations required for this
conversion are not currently exposed by the types generated by Candid
and didc, which means that each developer has to manually copy the data.
Long story short, it's not the best DX.

That's why, in our implementation of the Oisy standard, we discussed
providing an opinionated client library that can be used to interact
with Oisy, or wallet.

Given that most JavaScript developers are likely more familiar with
using the types exposed by ic-js to interact with well-known canisters
of the Internet Computer (IC), such as ledgers, rather than those
generated by Candid, this PR aims to expose the "converters" of
`ledger-icp` to allow the client library to obfuscate this complexity.

In short, this Oisy client library needs to be able to transform IC-js
types into Candid types the same way ic-js do. Therefore, to avoid code
duplication, this PR exposes the converters.

# Notes

Oisy signer lib is currently not open source as in development and not
yet reviewed by security but, happy to jump in a call to share more
details if anything is unclear.

# Changes

- Expose `@dfinity/ledger-icp` converters
- Generate test. For "consent message" I extracted a mock from existing
data as I knew this one existed since it was recently added.

Signed-off-by: David Dal Busco <david.dalbusco@dfinity.org>