Proposal based votable state (#725)
# Motivation

There are multiple reports from users who can’t vote. We believe the
root cause of this issue is the 100-ballot limit for neurons. If a user
has voted on more than 100 proposals (a likely scenario over a 4-day
period), they don’t receive ballots for proposals they’ve already voted
on, which remain technically votable due to this limit. The suggested
solution is to switch from using neuron ballots to proposal ballots when
calculating the votable state.

# Changes

- Use proposal ballots to get voting state.

# Tests

- Updated.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>