feat: install code arguments inherited from Candid (#768)
# Motivation

The `install_code` arguments for the `mode` have grown in complexity as
`Upgrade` now accepts parameters. Therefore, mapping the arguments those
not makes that much sense anymore as we would have to declare variant
within the types of the library as well. Therefore, this PR replaces the
custom params types with the inherited Candid types.

# Changes

- Replace `InstallMode` with Candid `canister_install_mode`
- Expose candid type
- Remove mapper
- Adapt code

---------

Signed-off-by: David Dal Busco <david.dalbusco@dfinity.org>
Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>