feat: extract ic-mgmt call transform (#774)
# Motivation

It’s literally impossible to test the internal `transform` function used
within the static `create` function of the `ic-mgmt` class. However, as
we need to address an issue to support WASM installation with chunks
(#773), we must extend it. That’s why this PR extracts the function into
a module, allowing us to provide tests for the mapping.

# Notes

We delegate the responsibility of calling the `transform` function to
`agent-js`, as it is used internally there and as we are "just"
implementing the interface exposed by agent-js. While it is not
thoroughly covered at the moment in the agent, I submitted a PR
[#954](https://github.com/dfinity/agent-js/pull/954) today that adds at
least one assertion.

No particular entry in CHANGELOG. This is an internal refactoring.

# Changes

- Extract internal `transform` function to a utility.
- JSdocs and tests.
- Improve types of the function (expected interface and return type)

---------

Signed-off-by: David Dal Busco <david.dalbusco@dfinity.org>
Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>