Update Candid files for nns and convert RefreshVotingPower (#779)
# Motivation

The automatic update PR (https://github.com/dfinity/ic-js/pull/778) had
failures in the `nns` and `cketh` packages.
This PR addresses the update of the `nns` package.

# Changes

1. Checked out `git checkout
release-2024-11-28_03-15-revert-hashes-in-blocks` in my `ic` repo.
2. Ran `scripts/import-candid ../../ic`
3. Ran `scripts/compile-idl-js`
4. Added conversion for `RefreshVotingPower` command.
5. Removed changes outside the `nns` package by running `git checkout
packages/{ckbtc,cketh,cmc,ledger-icp,ledger-icrc,sns}`

# Tests

The issue was caught by the compiler. There is no logic as the body of
the `RefreshVotingPower` type is empty. And there are no existing tests
for the changed function.

# Todos

- [x] Add entry to changelog (if necessary).