feat: add support for get_subnet_types_to_subnets in CMC (#782)
# Motivation

If a developer want to discover the list of subnet in which they can
create canister, they should actually query another endpoint of the CMC
in addition of the already exposed `get_default_subnets`.

The `get_subnet_types_to_subnets` provides additional subnets that can
be used to create canisters but, are not used by the CMC itself to
create canister on a random subnet when not targeted subnet ID is
provided.

A bit complicated but, once you know you need both, it's a bit more
clear.

# Changes

- Expose `get_subnet_types_to_subnets`.
- Expose did type for the response.

---------

Signed-off-by: David Dal Busco <david.dalbusco@dfinity.org>
Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>