feat: icrc map token metadata to record (#798)
# Motivation

I've been implementing the same mapper repeatedly, and today I needed it
in another library (the oisy-wallet-signer). Instead of reimplementing
it again, I added a utility to the ledger-icrc library to map the token
metadata information from a ledger response provided in form of Candid
arrays into a structured record.

# Changes

- Add utility `mapTokenMetadata`.

---------

Signed-off-by: David Dal Busco <david.dalbusco@dfinity.org>
Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>