Stop updating old ICRC1 index canister candid (#805)
# Motivation

The old ICRC1 index canister was deleted in
https://github.com/dfinity/ic/pull/3286
This caused the automatic candid update job to fail when it tried to
download the candid for this canister:

https://github.com/dfinity/ic-js/actions/runs/12626049412/job/35178612469

I think it will be good to keep the canister in `ic-js` for backwards
compatibility, at least for a while.
But since the canister code is deleted, the candid definitely won't be
changing anymore so we can stop automatically updating it.

# Changes

Stop updating the candid for the old ICRC1 index canister in
`scripts/import-candid`.

# Tests

Ran the script before and after and saw that it started working again.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary