Update old_list_neurons_service.certified.idl.js (#806)
# Motivation

We currently have an issue where fetching the neuron after a neuron
change (e.g., updating the dissolve delay) doesn’t return all new neuron
fields. In the UI, it appears as though the neuron activity fields were
updated (due to the optimistic substitution of the missing fields), but
they are not. The reason is that these requests use an old candid
version to support the older version of the ledger device. To retrieve
the new fields, we need to add them to the old list_neuron candid file
(as done here: https://github.com/dfinity/ic-js/pull/729).

# Changes

- Copy Neuron and NeuronInfo types from governance.certified.idl.js to
old_list_neurons_service.certified.idl.js.

# Tests

- Tested locally by upgrading ic-js to this version and verifying that
changing neurons (both normal and HW) returns the new fields and
displays the neuron activity status as expected.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.