feat: init new zod-schemas library (#812)
# Motivation

We have some Zod schemas that are repeated in OISY and
oisy-wallet-signer and are also used by the community—such as myself in
Juno. To avoid code duplication, it might be beneficial to create a
library for these schemas related to building apps on the IC.

# Notes

The package contains a dummy test to be removed as the CI requires at
least a test to complete.

# Changes

- Init library `@dfinity/zod-schemas` and tooling without any logic.
- Zod v3 as peer dependencies

---------

Signed-off-by: David Dal Busco <david.dalbusco@dfinity.org>
Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>