fix(security): expose certified flag for ICP to XDR conversion rate (#830)
# Motivation

It has been flagged as a security concern not to use a certified query
when reading the ICP to XDR conversion rate.

One option is to implement certificate validation, but its complexity
seems beyond the scope of the security ticket. Therefore, I am exposing
a new parameter to make the call an update call instead of a query call.

[SECFIND-421](https://dfinity.atlassian.net/browse/SECFIND-421)

# Changes

- New `certified` parameter for `getIcpToCyclesConversionRate` to use
the `certifiedService`

# Tests

- Updated existing test to check that `service` is call when `certified:
false`
- Add new test to check that `certifiedService` is call when `certified:
true`

# Todos

- [x] Add entry to changelog (if necessary).

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>