feat: Base64 string to and from Uint8Array (#828)
# Motivation

In today's Juno live session, I was looking to decode an `IcrcAccount`
provided by the OISY Wallet Signer to a string using the `owner` and
`subaccount`. The `subaccount`, in this case, is provided as a `base64`
string, so I needed to convert it to a `Uint8Array`. I could have done
this if I had access to a utility within the
`@dfinity/oisy-wallet-signer` library. However, this utility is not
exposed.

I considered exposing it in the signer library but later realized that
it makes more sense to move it to the `@dfinity/utils` package, as it
could be useful for other use cases.

# Changes

- Copy `base64.utils` from
https://github.com/dfinity/oisy-wallet-signer/blob/main/src/utils/base64.utils.ts
to ic-js

---------

Signed-off-by: David Dal Busco <david.dalbusco@dfinity.org>
Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>