feat: add single service for query and update calls (#871)
# Motivation

I would like to add @peterpeterparker 's code for the service util
`queryAndUpdate`. It was [added in
OISY](https://github.com/dfinity/oisy-wallet/blob/dfadb4a70e1b16644d613d931487f0a848d424e6/src/frontend/src/lib/actors/query.ic.ts#L35)
with the scope of making both type of calls to have a fast response
(`query`) with a consequent safer response (`update`), to avoid possible
information forging.

I think it should be part of the `ic-js` library since it is quite
useful, and it can come in handy in several other situations for the
developers, and for other repos.

The function accepts a request as input that can be executed both as
`query` and as `update` call. When executed, it handles the response
with callbacks, including errors.

Furthermore, it provides the possibility to:
- make only one type of call, `query` or `update`, if necessary;
- have a specific callback for `update` calls;
- decide if both calls must be awaited or it can resolves after the
first one resolves.

# Changes

- Add types for the new service.
- Copy the function `queryAndUpdate` from the OISY repo and paste it in
`@dfinity/utils` package.
- Export the module and the types in the `index` file.

# Tests

Added extensive tests to cover all possible scenarios, even an unlikely
one where the `update` call is faster than the `query`.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>