chore: eslint v9 with eslint-config-oisy-wallet (#895)
# Motivation

We are using an outdated version of ESLint and have to migrate to ESLint
v9, as we did in our repo and apps.

While slightly opinionated, but fundamentally matching our code style, I
propose to migrate using the `@dfinity/eslint-config-oisy-wallet`, which
we are using in OISY and various repos of the foundation.

This will ease the maintenance and reduce the related amount of work in
the future while providing a set of rules anybody can contribute to.

# Changes

- Remove eslint v8
- Install and configure `@dfinity/eslint-config-oisy-wallet`
- Disable all rules that do not comply. Those will have to be fixed
iteratively in separate PR
- `allowJs` in `tsconfig.eslint.json` to lint `eslint.config.mjs`
- Update lint rule in `debounce.utils`. The rule it was using does not
exist anymore hence had to be fixed in this PR.
- Build utils before running linter in CI because
`eslint-config-oisy-wallet` uses `non|isNullish` in custom rules.

---------

Signed-off-by: David Dal Busco <david.dalbusco@dfinity.org>
Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>