fix: assert checksum when parsing ICP AccountIdentifier from Hex (#948)
# Motivation

@ulan noticed that the ICP `AccountIdentifier.fromHex` function was not
asserting the checksum of the provided input.
In other words, it was possible to generate an account identifier object
representing an invalid account. That’s why he opened a PR to add such a
check, as it's safer to enforce validation at the class level.

It's worth noting that this is a potentially breaking change, since
applications that previously accepted invalid account identifiers will
now fail. However, we can argue that those invalid accounts were already
broken by definition.

# Notes

This PR includes Ulan's original code along with a few minor adjustments
based on my review comments.
Since we don't currently accept external contributions in this repo, we
agreed that I would submit the combined version.

# Changes

- Add an assertion in `AccountIdentifier.fromHex` to ensure the checksum
of the provided hex string is valid.

---------

Signed-off-by: David Dal Busco <david.dalbusco@dfinity.org>
Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>