feat: migrate to zod/v4 (#951)
# Motivation

We want to upgrade to Zod v4 and also try to leverage the new mini
version to reduce bundle size.

# Zod-mini

I originally started using Zod v4-mini but then read the author of Zod
is advicing to commonly stick to the normal version of Zod.
https://zod.dev/packages/mini?id=when-not-to-use-zod-mini

As a result, I went for v4 because in addition it requires a bit less
changes as the `refine` pattern must not be migrated to `check`.

Note that this PR also convert the deprecated `message`. Did not had too
but, since it's a small library, I thought to do it once for all.

# Sources

- https://zod.dev/v4
- https://zod.dev/packages/mini

# Changes

- Update `zod-schemas` according migration guide.

---------

Signed-off-by: David Dal Busco <david.dalbusco@dfinity.org>
Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>