build: add explicit lib requirement on buffer to nns (#963)
# Motivation

`@dfinity/nns` has unfortunately a requirement on the `buffer` library,
which is really a pity. This is notably used to convert identifier from
and to bytes back and forth. I guess it's historical code of the
governance and/or the hardware wallet.

Anyway, on the other end, agent-js removes its dependency to `buffer` in
its version v3. Depency which is currently inherited to build the
library.

That is why - because I think that improving nns would require a bit of
time - I propose to "just" set `buffer` as an explicit dependency of
`nns-js`.

# Changes

- ` npm i buffer --save-peer -w packages/nns`

---------

Signed-off-by: David Dal Busco <david.dalbusco@dfinity.org>
Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>