test: remove `.only` modifier for tests of util `smallerVersion` (#970)
# Motivation

Since we are going to migrate to Vitest, we receive an error while
testing:

```css
→ [Vitest] Unexpected .only modifier. Remove it or pass --allowOnly argument to bypass this error
```

meaning that Vitest blocks `.only` unless we explicitly allow it.

This is a safety feature to prevent accidentally committing focused
tests (`it.only`, `describe.only`) which would skip the rest of your
suite.

However, the test suite specific for util `smallerVersion` works
correctly, so we remove the modifier.