feat: replace randombytes by crypto.getRandomValue in nns-js (#962)
# Motivation

I'm not sure why it was originally introduced in #44 (@lmuntaner) can
you double check?), but the use of the randombytes library in NNS-js no
longer seems necessary. A random nonce can be generated using the
standard
[crypto.getRandomValues](https://developer.mozilla.org/en-US/docs/Web/API/Crypto/getRandomValues),
which has been supported in both browsers and Node.js for quite some
time.

Furthermore, Agent-js v3 (finally) removes its dependency on `buffer`,
which was also required by the `randombytes` library (see #962). This
means we would either need to reintroduce `buffer` as a dependency in
NNS-js — which we obviously want to avoid — or find an alternative
approach.

# Changes

- Use `const nonceBytes = crypto.getRandomValues(new Uint8Array(8));`
for the feature
- Adapt mock for test

---------

Signed-off-by: David Dal Busco <david.dalbusco@dfinity.org>
Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>