feat: replace usage of Buffer with Uint8Array in checkAccountId (#1011)
# Motivation

On one hand, we generally want to move away from using `Buffer`, on the
other, when bumping TypeScript in #1004, the compiler reports issues
related to its usage.

This PR replaces `Buffer` with `Uint8Array` in the `checkAccountId`
function of `@dfinity/ledger-icp` to address both concerns.

# Changes

- Use existing utils and new comparison utils (introduced in #1005) to
replace buffer with Uint8Array.
- Improve tests to validate checksum.

---------

Signed-off-by: David Dal Busco <david.dalbusco@dfinity.org>